requirements.txt
astroid==2.12.10
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.1
dill==0.3.5.1
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pylint==2.15.3
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
python-gitlab==3.10.0
requests==2.27.1
requests-toolbelt==0.9.1
tomli==2.0.1
tomlkit==0.11.5
urllib3==1.26.8
wrapt==1.14.1
